+++ write_script_multi.py
@@ -491,8 +491,8 @@ try:
 	sampler.reset()
 
 	t_org = time.time()
-	print("total iterations left: ", (tot_iter)/100)
-	for i in range(0, (tot_iter)/100):
+	print("total iterations left: ", int((tot_iter)/100))
+	for i in range(0, int((tot_iter)/100)):
 		print('Starting', i)
 		t0 = time.time()
 		pos, prob, state = sampler.run_mcmc(pos, 100, lnprob0 = prob, rstate0 = state)
